refactor: split OCR utilities
